main_server/requirements.txt
websockets
GPIO
mpu6050-raspberrypi
python3-smbus